Remove ATL dependency from Mock Reference Tracker runtime library.
