+++ src/coreclr/tests/src/Interop/common/ComHelpers.h
@@ -331,3 +331,69 @@ public: // IUnknown
 
     DEFINE_REF_COUNTING();
 };
+
+template<typename T>
+struct ComSmartPtr
+{
+    T* p;
+
+    ComSmartPtr()
+        : p{ nullptr }
+    { }
+
+    ComSmartPtr(_In_ T* t)
+        : p{ t }
+    {
+        if (p != nullptr)
+            (void)p->AddRef();
+    }
+
+    ComSmartPtr(_In_ const ComSmartPtr&) = delete;
+
+    ComSmartPtr(_Inout_ ComSmartPtr&& other)
+        : p{ other.Detach() }
+    { }
+
+    ~ComSmartPtr()
+    {
+        if (p != nullptr)
+            (void)p->Release();
+    }
+
+    ComSmartPtr& operator=(_In_ const ComSmartPtr&) = delete;
+
+    ComSmartPtr& operator=(_Inout_ ComSmartPtr&& other)
+    {
+        Attach(other.Detach());
+        return (*this);
+    }
+
+    operator T*()
+    {
+        return p;
+    }
+
+    T** operator&()
+    {
+        return &p;
+    }
+
+    T* operator->()
+    {
+        return p;
+    }
+
+    void Attach(_In_opt_ T* t) noexcept
+    {
+        if (p != nullptr)
+            (void)p->Release();
+        p = t;
+    }
+
+    T* Detach() noexcept
+    {
+        T* tmp = p;
+        p = nullptr;
+        return tmp;
+    }
+};
